Refactored getTransactionSummary endpoint, abstracted getDateRange functionality to transactions helper method
