+++ expense-tracker-backend/src/helpers/transactions/transactionControllerHelpers.js
@@ -1,3 +1,4 @@
+import { getCurrentMonthDateRange, getMonthDateRange, getCurrentYearDateRange, validateMonthYear } from './transactionSummaryHelpers.js';
 /**
  * Helper functions for transaction controllers
  */
@@ -16,4 +17,20 @@ export const formatTransactionsResponse = (transactions, lastEvaluatedKey) => {
       lastEvaluatedKey: lastEvaluatedKey ? JSON.stringify(lastEvaluatedKey) : null
     }
   };
+};
+
+export const getDateRange = (period, month, year) => {
+    let dateRange;
+    if (period === 'current-year') {
+      dateRange = getCurrentYearDateRange();
+    } else if (month && year) {
+      const validationErrors = validateMonthYear(parseInt(month), parseInt(year));
+      if (validationErrors) {
+        throw { type: 'validation', message: 'Invalid month/year', details: validationErrors };
+      }
+      dateRange = getMonthDateRange(parseInt(month), parseInt(year));
+    } else {
+      dateRange = getCurrentMonthDateRange();
+    }
+    return dateRange;
 };
